// src/generated/dbTypes.ts
/*
* This file was generated by a tool.
* Rerun sql-ts to regenerate this file.
*/
export interface PlaylistItemEntity {
  'playlist_id': number;
  'pos': number;
  'track_id': number;
}
export interface PlaylistEntity {
  'all_items'?: any | null;
  'description'?: any | null;
  'distinguished_kind'?: any | null;
  'folder'?: any | null;
  'music'?: any | null;
  'name'?: string | null;
  'parent_persistent_id'?: any | null;
  'playlist_id'?: number | null;
  'playlist_persistent_id'?: string | null;
  'smart_criteria'?: any | null;
  'smart_info'?: any | null;
}
export interface TrackEntity {
  'album'?: string | null;
  'album_artist'?: string | null;
  'album_rating'?: any | null;
  'album_rating_computed'?: any | null;
  'artist'?: string | null;
  'artwork_count'?: any | null;
  'bit_rate'?: number | null;
  'bpm'?: number | null;
  'clean'?: any | null;
  'comments'?: string | null;
  'compilation'?: number | null;
  'composer'?: string | null;
  'date_added'?: string | null;
  'date_modified'?: string | null;
  'disabled'?: number | null;
  'disc_count'?: number | null;
  'disc_number'?: number | null;
  'explicit'?: any | null;
  'file_folder_count'?: any | null;
  'file_type'?: string | null;
  'genre'?: string | null;
  'grouping'?: any | null;
  'has_video'?: any | null;
  'hd'?: any | null;
  'kind'?: string | null;
  'library_folder_count'?: any | null;
  'location'?: string | null;
  'movement_count'?: any | null;
  'movement_name'?: any | null;
  'movement_number'?: any | null;
  'music_video'?: any | null;
  'name'?: string | null;
  'normalization'?: number | null;
  'part_of_gapless_album'?: any | null;
  'persistent_id'?: string | null;
  'play_count'?: number | null;
  'play_date'?: number | null;
  'play_date_utc'?: string | null;
  'protected'?: any | null;
  'purchased'?: any | null;
  'rating'?: number | null;
  'release_date'?: any | null;
  'sample_rate'?: number | null;
  'series'?: any | null;
  'size'?: number | null;
  'skip_count'?: any | null;
  'skip_date'?: any | null;
  'sort_album'?: any | null;
  'sort_album_artist'?: any | null;
  'sort_artist'?: any | null;
  'sort_composer'?: any | null;
  'sort_name'?: any | null;
  'start_time'?: number | null;
  'stop_time'?: number | null;
  'total_time'?: number | null;
  'track_count'?: number | null;
  'track_id'?: number;
  'track_number'?: number | null;
  'track_type'?: string | null;
  'video_height'?: any | null;
  'video_width'?: any | null;
  'volume_adjustment'?: number | null;
  'work'?: any | null;
  'year'?: number | null;
}
export interface VoxAlbumEntity {
  'album'?: string | null;
  'artist'?: string | null;
  'id'?: number | null;
  'sentence'?: string | null;
}
export interface VoxArtistEntity {
  'artist'?: string | null;
  'id'?: number | null;
  'sentence'?: string | null;
}
export interface VoxGenreEntity {
  'genre'?: string | null;
  'id'?: number | null;
  'sentence'?: string | null;
}
export interface VoxPlaylistEntity {
  'id'?: number | null;
  'playlist_id'?: number | null;
  'sentence'?: string | null;
}
export interface VoxTrackEntity {
  'id'?: number | null;
  'sentence'?: string | null;
  'track_id'?: number | null;
}